Basic Real Estate Simulation
Developed using C# as part of the Programming Language-3 course, the Real Estate Registration Project includes a user login screen, property listing, and house addition functionalities. Usernames and passwords are stored in a .txt file. Upon successful login, the user is directed to an interface listing
features such as price, rent, deposit, floor number, and room count for properties available for sale or rent. This interface includes a button to add a new house and another to update the list. When the house addition button is clicked, a form appears where users enter house details. If negative values are entered, an error message is displayed, and incorrect entries are logged in a file. After adding a property, the "Update List" button ensures the newly added
house appears in the table.